Add files via upload
We are working for an e-learning company that offers courses on programming. Most of the courses are on web and mobile development, but we also cover many other domains, like data science, game development... We want to promote our product and we'd like to invest some money in advertisement. Our goal in this project is to find out the 2 best markets to advertise our product in.

The dataset is from freeCodeCamp's 2017 New Coder Survey. We can find the data here: https://github.com/freeCodeCamp/2017-new-coder-survey There are some statistic from the survey that we can refer to the link: https://www.freecodecamp.org/news/we-asked-20-000-people-who-they-are-and-how-theyre-learning-to-code-fff5d668969/